refactor(condvar): clone task wakers while registering
A normal task waker clone is an inexpensive reference-count update, so the waiter lock does not need a clone-and-recheck loop. Replaced and cancelled wakers still leave the lock before destruction; retain one focused regression test for that observable deadlock boundary and remove the clone-callback test suite.
